[APPLE-1243] Calling script to generate IP address.
If we're in the Scheme called
"ParkinsonUsingDataVerificationServer" --
which lets us log Sage data to our
dataVerificationServer -- and we want to run
the app from a phone, we need to know what IP
address the server is running on.  This commit
makes Parkinson's call the script I wrote which
generates that IP address (for situations where
you're running the server on your laptop).

The link to AppCore ties this commit to the
AppCore commit containing a version of that
script which only runs when that special
Scheme is being used.  (Previous versions of
the script always ran, regardless of what Scheme
was in use.)