[APPLE-1243] Documentation for testing passive data collection.
Added comments to Dhanush's sample code
for invoking the passive-data collectors.

In person, Dhanush told me how to use this,
including how to trigger the uploads quickly
(set the "threshhold" to 100 bytes).  I added
the lines of code which do that, then set up the
formatting and comments in a way to make that
(hopefully) easy to read.  Then I commented out
the whole thing, as Dhanush had left it.