[APPLE-1697] App-level Test cases for the MedTracker CoreData stuff.
Created a test case up here in the Parkinson's main app for the CoreData storage features I'm building for the Medication Tracker.

We have to run the tests in the main app because, otherwise, we don't have access to the CoreData infrastructure features (the APCDataSubstrate) which get instantiated up here.

This commit requires the features in the linked version of AppCore.  (That AppCore build is completely independent of this commit, though.)