[APPLE-1697] Retrieving all data (currently empty lists).
Wrote the bare-minimum fetch requests to pull all my data from CoreData.

Interesting note:  when I'm doing these tests, the data I create on the Simulator in one run of the test case PERSISTS, across runs.  For my purposes, that's GREAT, because it gives me an easy and reliable way to test.  It's also something to watch for when we do these test cases.  Per Farhan:  to prevent that, I can use the -tearDown method to delete that stuff.