[APPLE-1697] Wrote test case to delete all data.
Purposely chose NOT to put this into the main code base; it's fundamentally dangerous.  But it's also fundamentally necessary during testing, particularly because I want to be sure my "update" concepts work:  when we re-load colors from a prefs file (so, like, we can modify the colors and then re-ship the apps), I want to update the existing colors, not delete them or add colors with duplicate names.