[APPLE-1997] Applying ViewController changes for RK522.
Under Ed's guidance:  in three ViewControllers, removed two outdated methods, replacing them with the current, more-general method.

Also logging an error if we get one, and the "result" enum value says "failed."

The result compiles and runs.

This very much depends on the linked version of AppCore, which contains the new ResearchKit.